get integer cast operators their own header
